pyproject.toml
[project]
name = "fizzbuzz-bench"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.75.0",
    "google-genai>=1.0.0",
    "ipdb>=0.13.13",
    "openai>=2.11.0",
    "together>=1.5.31",
    "tqdm>=4.67.1",
]
